+++ CHANGELOG.md
@@ -14,6 +14,7 @@ September 2025
   - draw symbol outline
   - adjust icon and label scale
   - render symbol amplifiers
+  - nested folders for sides and groups
 - Units can now be added directly to a side. A side group is no longer required.
 
 May 2025